# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BPUT Feedback Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Department {
  id: number;
  code: string;
  name: string;
  totalFaculty: number;
  totalCourses: number;
  avgRating?: number;
}

export interface Faculty {
  id: number;
  name: string;
  email?: string;
  designation: string;
  departmentId: number;
  departmentName?: string;
  avgRating?: number;
  totalFeedbackCount?: number;
}

export interface CreateFacultyBody {
  name: string;
  email?: string;
  designation: string;
  departmentId: number;
}

export interface Course {
  id: number;
  code: string;
  name: string;
  departmentId: number;
  departmentName?: string;
  facultyId?: number;
  facultyName?: string;
  semester: number;
  academicYear: string;
  credits?: number;
  avgRating?: number;
  feedbackCount?: number;
}

export interface CreateCourseBody {
  code: string;
  name: string;
  departmentId: number;
  facultyId?: number;
  semester: number;
  academicYear: string;
  credits?: number;
}

export type FeedbackEntryFeedbackType =
  (typeof FeedbackEntryFeedbackType)[keyof typeof FeedbackEntryFeedbackType];

export const FeedbackEntryFeedbackType = {
  semester_end: "semester_end",
  mid_semester: "mid_semester",
  event_based: "event_based",
  placement: "placement",
} as const;

export interface FeedbackEntry {
  id: number;
  referenceId: string;
  courseId: number;
  courseName?: string;
  courseCode?: string;
  facultyId?: number;
  facultyName?: string;
  departmentId: number;
  departmentName?: string;
  semester: number;
  academicYear: string;
  studentYear?: number;
  section?: string;
  feedbackType: FeedbackEntryFeedbackType;
  /**
   * @minimum 1
   * @maximum 5
   */
  ratingCourseContent: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  ratingTeachingQuality: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  ratingLabFacilities: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  ratingStudyMaterial: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  ratingOverall: number;
  comments?: string;
  isAnonymous: boolean;
  createdAt: string;
}

export type SubmitFeedbackBodyFeedbackType =
  (typeof SubmitFeedbackBodyFeedbackType)[keyof typeof SubmitFeedbackBodyFeedbackType];

export const SubmitFeedbackBodyFeedbackType = {
  semester_end: "semester_end",
  mid_semester: "mid_semester",
  event_based: "event_based",
  placement: "placement",
} as const;

export interface SubmitFeedbackBody {
  courseId: number;
  facultyId?: number;
  studentYear?: number;
  section?: string;
  feedbackType: SubmitFeedbackBodyFeedbackType;
  /**
   * @minimum 1
   * @maximum 5
   */
  ratingCourseContent: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  ratingTeachingQuality: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  ratingLabFacilities: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  ratingStudyMaterial: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  ratingOverall: number;
  comments?: string;
  isAnonymous?: boolean;
}

export type FeedbackWindowFeedbackType =
  (typeof FeedbackWindowFeedbackType)[keyof typeof FeedbackWindowFeedbackType];

export const FeedbackWindowFeedbackType = {
  semester_end: "semester_end",
  mid_semester: "mid_semester",
  event_based: "event_based",
  placement: "placement",
} as const;

export interface FeedbackWindow {
  id: number;
  title: string;
  feedbackType: FeedbackWindowFeedbackType;
  academicYear: string;
  semester: number;
  startDate: string;
  endDate: string;
  isActive: boolean;
  departmentIds?: number[];
  createdAt: string;
}

export type CreateWindowBodyFeedbackType =
  (typeof CreateWindowBodyFeedbackType)[keyof typeof CreateWindowBodyFeedbackType];

export const CreateWindowBodyFeedbackType = {
  semester_end: "semester_end",
  mid_semester: "mid_semester",
  event_based: "event_based",
  placement: "placement",
} as const;

export interface CreateWindowBody {
  title: string;
  feedbackType: CreateWindowBodyFeedbackType;
  academicYear: string;
  semester: number;
  startDate: string;
  endDate: string;
  departmentIds?: number[];
}

export interface UpdateWindowBody {
  title?: string;
  startDate?: string;
  endDate?: string;
  isActive?: boolean;
}

export type DashboardSummaryRatingBreakdown = {
  courseContent: number;
  teachingQuality: number;
  labFacilities: number;
  studyMaterial: number;
  overall: number;
};

export interface DashboardSummary {
  totalFeedback: number;
  totalCourses: number;
  totalFaculty: number;
  totalDepartments: number;
  avgOverallRating: number;
  activeWindows: number;
  feedbackThisMonth: number;
  topDepartment?: string;
  recentFeedback: FeedbackEntry[];
  ratingBreakdown: DashboardSummaryRatingBreakdown;
}

export interface DepartmentAnalytics {
  departmentId: number;
  departmentName: string;
  totalFeedback: number;
  avgRating: number;
  avgCourseContent: number;
  avgTeachingQuality: number;
  avgLabFacilities: number;
  avgStudyMaterial: number;
  topCourses: Course[];
  facultyPerformance: Faculty[];
}

export interface FacultyAnalytics {
  facultyId: number;
  facultyName: string;
  designation?: string;
  departmentName?: string;
  totalFeedback: number;
  avgOverall: number;
  avgCourseContent: number;
  avgTeachingQuality: number;
  avgLabFacilities: number;
  avgStudyMaterial: number;
  courses: Course[];
  recentComments: string[];
}

export interface TrendPoint {
  academicYear: string;
  semester: number;
  avgRating: number;
  feedbackCount: number;
  label: string;
}

export interface TopRated {
  topFaculty: Faculty[];
  topCourses: Course[];
}

export type ListFacultyParams = {
  departmentId?: number;
};

export type ListCoursesParams = {
  departmentId?: number;
  semester?: number;
  academicYear?: string;
};

export type ListFeedbackParams = {
  courseId?: number;
  departmentId?: number;
  facultyId?: number;
  semester?: number;
  academicYear?: string;
  feedbackType?: ListFeedbackFeedbackType;
};

export type ListFeedbackFeedbackType =
  (typeof ListFeedbackFeedbackType)[keyof typeof ListFeedbackFeedbackType];

export const ListFeedbackFeedbackType = {
  semester_end: "semester_end",
  mid_semester: "mid_semester",
  event_based: "event_based",
  placement: "placement",
} as const;

export type GetDepartmentAnalyticsParams = {
  academicYear?: string;
};

export type GetTrendsParams = {
  departmentId?: number;
  metric?: GetTrendsMetric;
};

export type GetTrendsMetric =
  (typeof GetTrendsMetric)[keyof typeof GetTrendsMetric];

export const GetTrendsMetric = {
  overall: "overall",
  course_content: "course_content",
  teaching_quality: "teaching_quality",
  lab_facilities: "lab_facilities",
  study_material: "study_material",
} as const;