# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BPUT Feedback Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List courses
      parameters:
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: semester
          in: query
          schema:
            type: integer
        - name: academicYear
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "201":
          description: Created course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"

  # ── Feedback ───────────────────────────────────────────────────────────────
...
  /analytics/department/{departmentId}:
    get:
      operationId: getDepartmentAnalytics
      tags: [analytics]
      summary: Department-wise analytics
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: integer
        - name: academicYear
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Department analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartmentAnalytics"
  /analytics/faculty/{facultyId}:
    get:
      operationId: getFacultyAnalytics
      tags: [analytics]
      summary: Faculty performance analytics
      parameters:
        - name: facultyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Faculty analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FacultyAnalytics"
  /analytics/trends:
    get:
      operationId: getTrends
      tags: [analytics]
      summary: Semester-wise trend data
      parameters:
        - name: departmentId
          in: query
          schema:
            type: integer
        - name: metric
          in: query
          schema:
            type: string
            enum: [overall, course_content, teaching_quality, lab_facilities, study_material]
      responses:
        "200":
          description: Trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendPoint"
...
components:
  schemas:
...
    Faculty:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        designation:
          type: string
        departmentId:
          type: integer
        departmentName:
          type: string
        avgRating:
          type: number
          format: float
        totalFeedbackCount:
          type: integer
      required: [id, name, designation, departmentId]
...
    Course:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        departmentId:
          type: integer
        departmentName:
          type: string
        facultyId:
          type: integer
        facultyName:
          type: string
        semester:
          type: integer
        academicYear:
          type: string
        credits:
          type: integer
        avgRating:
          type: number
          format: float
        feedbackCount:
          type: integer
      required: [id, code, name, departmentId, semester, academicYear]
    CreateCourseBody:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        departmentId:
          type: integer
        facultyId:
          type: integer
        semester:
          type: integer
        academicYear:
          type: string
        credits:
          type: integer
      required: [code, name, departmentId, semester, academicYear]
...
    DepartmentAnalytics:
      type: object
      properties:
        departmentId:
          type: integer
        departmentName:
          type: string
        totalFeedback:
          type: integer
        avgRating:
          type: number
          format: float
        avgCourseContent:
          type: number
          format: float
        avgTeachingQuality:
          type: number
          format: float
        avgLabFacilities:
          type: number
          format: float
        avgStudyMaterial:
          type: number
          format: float
        topCourses:
          type: array
          items:
            $ref: "#/components/schemas/Course"
        facultyPerformance:
          type: array
          items:
            $ref: "#/components/schemas/Faculty"
      required:
        - departmentId
        - departmentName
        - totalFeedback
        - avgRating
        - avgCourseContent
        - avgTeachingQuality
        - avgLabFacilities
        - avgStudyMaterial
        - topCourses
        - facultyPerformance
    FacultyAnalytics:
      type: object
      properties:
        facultyId:
          type: integer
        facultyName:
          type: string
        designation:
          type: string
        departmentName:
          type: string
        totalFeedback:
          type: integer
        avgOverall:
          type: number
          format: float
        avgCourseContent:
          type: number
          format: float
        avgTeachingQuality:
          type: number
          format: float
        avgLabFacilities:
          type: number
          format: float
        avgStudyMaterial:
          type: number
          format: float
        courses:
          type: array
          items:
            $ref: "#/components/schemas/Course"
        recentComments:
          type: array
          items:
            type: string
      required:
        - facultyId
        - facultyName
        - totalFeedback
        - avgOverall
        - avgCourseContent
        - avgTeachingQuality
        - avgLabFacilities
        - avgStudyMaterial
        - courses
        - recentComments
    TrendPoint:
      type: object
      properties:
        academicYear:
          type: string
        semester:
          type: integer
        avgRating:
          type: number
          format: float
        feedbackCount:
          type: integer
        label:
          type: string
      required: [academicYear, semester, avgRating, feedbackCount, label]
...