# lib/api-zod/src/generated/types/course.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BPUT Feedback Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface Course {
  id: number;
  code: string;
  name: string;
  departmentId: number;
  departmentName?: string;
  facultyId?: number;
  facultyName?: string;
  semester: number;
  academicYear: string;
  credits?: number;
  avgRating?: number;
  feedbackCount?: number;
}