# lib/api-zod/src/generated/types/faculty.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BPUT Feedback Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface Faculty {
  id: number;
  name: string;
  email?: string;
  designation: string;
  departmentId: number;
  departmentName?: string;
  avgRating?: number;
  totalFeedbackCount?: number;
}