# lib/api-zod/src/generated/types/trendPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BPUT Feedback Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface TrendPoint {
  academicYear: string;
  semester: number;
  avgRating: number;
  feedbackCount: number;
  label: string;
}